+++ Modules/Settings/SettingsDefaults.lua
@@ -52,6 +52,7 @@ function DP_SettingsDefaults:Load()
         sessionIgnoredItems = {
         },
         onlySoulbound = false,
+        showItemQuality = false,
       },
     },
   }